feat(playerProfile): add tactical role assignment
Extract ROLE_OPTIONS_BY_POSITION to shared lib/playerRoles.ts and add
a Tactical Role select in the player profile for own-club players; also
fix squad contract column i18n key (playerProfile.contract → common.contract).
